git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/trunk@3312 a63bd714-7e14-4b5e-94e7-302c8c8ff188

+++ src/particleContainer/ParticleContainer.cpp
@@ -31,7 +31,7 @@ double ParticleContainer::getBoundingBoxMin(int dimension) const {
 double ParticleContainer::getBoundingBoxMax(int dimension) const {
 	return this->_boundingBoxMax[dimension];
 }
-double ParticleContainer::getHaloWidthNumCells() {
+int ParticleContainer::getHaloWidthNumCells() {
 	return 0;
 }
 void ParticleContainer::updateMoleculeCaches() {
